story-enhancer-ai/RAG_ENV/requirements.txt
# === CORE === 
python-dotenv==1.0.1
pydantic==2.7.1
typing-extensions==4.9.0
httpx==0.27.0
aiofiles==23.2.1
tenacity==8.3.0
tqdm==4.66.4
psutil==5.9.8

# === LANGCHAIN + RAG ===
langchain
langchain-community
langchain-openai
chromadb==0.4.24
sentence-transformers
openai==1.35.3
numpy==1.26.3
pandas==2.2.2

# === DOCUMENT PROCESSING ===
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
unstructured==0.13.7
markdown==3.6
python-multipart==0.0.9

# === FRONTEND (Opcional) ===
streamlit==1.34.0
plotly==5.20.0

# === DEV TOOLS (Opcional) ===
pytest==8.2.1
black==24.4.2
flake8==7.0.0
mypy==1.10.0
